requirements.txt
aiogram==3.4.1
pyrogram==2.0.106
tgcrypto==1.2.5
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
asyncpg==0.29.0
python-dotenv==1.0.0
httpx==0.26.0
jinja2==3.1.3


